Free GPT+ Using gemini/requirements.txt
streamlit
python-dotenv
google.generativeai
langchain
chromadb
PyPDF2